pyproject.toml
[project]
name = "interview"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "matplotlib",
    "rerun-sdk",
    "opencv-python",
    "tqdm",
    "pyyaml>=6.0.3",
]
